moved handlers to conroller folder, commented out mongodb from server since not using yet and causing errors, turned interactions into objects with keys and values. Need to handle errors and return results not only as a console.log
