+++ lib/sendgrid.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const carrierHandler = require('../lib/carrierhandler');
+const carrierHandler = require('../controller/carrierhandler');
 
 module.exports = exports = function(phoneNumber, carrier, drug){
   let helper = require('sendgrid').mail,
